+++ src/components/AccountManagement/DeleteUsers/DeleteUsers.js
@@ -4,7 +4,7 @@ import { DataGrid } from "@material-ui/data-grid";
 import Axios from "axios";
 import { Button } from "@material-ui/core";
 import { Grid } from "@material-ui/core";
-import UserContext from "../../context/UserContext";
+import UserContext from "../../../context/UserContext";
 
 // Structure of the User's data
 const columns = [
