+++ src/components/AccountManagement/Register/Register.js
@@ -1,8 +1,8 @@
 import React from "react";
 import "./Register.css";
 import { Grid } from "@material-ui/core";
-import ImageDynamic from "../CovidStatsImg/ImageDynamic";
-import RegisterForm from "./RegisterForm";
+import ImageDynamic from "../../CovidStatsImg/ImageDynamic";
+import RegisterForm from "./Components/RegisterForm";
 
 export default function Register() {
   return (
