+++ src/components/PasswordChange.js
@@ -1,10 +1,10 @@
 import React, { useState, useContext } from "react";
-import "../pages/profile.css";
+import "./pages/profile.css";
 import Button from "@material-ui/core/Button";
-import SuccessAlert from "./SuccessAlert";
-import ErrorAlert from "./ErrorAlert";
+import SuccessAlert from "./AccountManagement/Register/Components/SuccessAlert";
+import ErrorAlert from "./AccountManagement/Register/Components/ErrorAlert";
 
-import UserContext from "../../context/UserContext";
+import UserContext from "../context/UserContext";
 import Axios from "axios";
 import Grid from "@material-ui/core/Grid";
